Created project structure for part 2.
This time I'm using two binaries, one for testing and one for main
so that main looks squeaky clean by the end of it all.

I'm currenty manually creating the input file. I can't seem to
figure out how to get cat to not emit a newline. The newline ends
up at the end of the input before the closing )"; for the raw string,
so the input has a trailing newline. This is kind of annoying because
without that newline I can treat newlines as a special character,
but when it's present I have to be more general in how I handle them.