I'm relatively confident in the correctness of my solution for #7, but the input length is so large that it starts  to really strain to handle such large variadic templates, I think. I may try to hook up a profiler and see where it's spending most of its time.
Another option is to chunk the list into N sections and handle
each one as a sub problem. It's definitely possible to implement
it that way, but it would be fairly complex. I'm not certain that
it would actually improve things, so I'm discouraged from putting
in the time to try.